+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Features
 - configurable music switch delay when changing between exterior cells.
 - option to carry over the region of the last exterior cell to an interior cell.
+- enemyFaction filter for soundbanks to restrict combat music for specific factions.
 
 ## Bugfixes
 - fixed cellNamePatternsExclude filter in soundbanks not working.